Refactor CharacterHooks into ActorModule.
Collate the two conflicting ActorModule hooks.
